Fill OUTPUT_DIRS with rule for libfoo.so linkage
Output directory, that stay after '|' in .so linkage rule
should be mentioned in OUTPUT_DIRS. Otherwise, in case of make
jobs parallelization has place, the rule for directory creation
will be missed and job for .so confuse:

  make -j4 --debug=j DEBUG_RULES=1
  GNU Make 3.82
  Built for x86_64-unknown-linux-gnu
  Copyright (C) 2010  Free Software Foundation, Inc.
  License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>
  This is free software: you are free to change and redistribute it.
  There is NO WARRANTY, to the extent permitted by law.
  Need a job token; we don't have children
  ../../gmake/top.mak:113: obj/gcc/so/:  ()
  Putting child 0x13d9270 (obj/gcc/so/) PID 5693 on the chain.
  make: *** No rule to make target `obj/gcc/so', needed by `obj/gcc/so/libtest.so.0.1.2'.  Stop.
  make: *** Waiting for unfinished jobs....
  Live child 0x13d9270 (obj/gcc/so/) PID 5693
  Reaping winning child 0x13d9270 PID 5693
  Removing child 0x13d9270 PID 5693 from chain.

You see that rule for obj/gcc/so present, but in job different from
job with rule for obj/gcc/so/libtest.so.0.1.2. The obj/gcc/so/ added
in OUTPUT_DIRS, but near the rule for .o (compilation) that placed
in another make job.